requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.3.1
cx-Oracle==8.1.0
Django==3.1.4
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==3.0.0a1
MarkupSafe==2.0.0a1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
ulid-py==1.1.0
uritemplate==3.0.1
urllib3==1.25.10
pandas==1.2.4
numpy==1.20.3
